VSE: Set "box select" default tool for preview
This has been discussed several times in chat, but no resolution was
determined, so I'm creating this PR to get feedback on this change.

Current default tool is "sample" which is a bit odd, and coupled with
the fact that the toolbar is hidden by default (separate issue which
could be addressed in another PR) this can be confusing to new users,
leading them to believe tweak/select tools are not supported for the VSE
preview.

I haven't looked into any versioning code w/ this proposal since I
figure it's not necessary to change existing projects (a default only
really applies to new projects).