Fix broken tests due to side effects in realize_instances.
The changed line here should not have an effect because it combines two
booleans, but using the boolean `||` operator instead of bitwise `|`
does not evaluated the second term if the first one is true.
The side effects are needed, but should be better explained. This will
be fixed in a separate PR.