Refactor: Sculpt: Require object for position access
Part of #118145.
In preparation for removing the duplicated position arrays from the
pbvh::Tree structure, change the API functions that access the arrays
to require object arguments. This will allow retrieving data from the
original mesh or the evaluated deform mesh as necessary.

The pbvh code isn't really the right place for this, but neither is
the sculpt or paint code in my opinion, since this concept is not specific to sculpting or painting. For now keep the API in the same
place. It probably makes sense to move a bunch of these functions
at some point anyway.