Removed the "None" constraint type.
It's already possible to disable constraints, so a None type is
redundant.

Disallowing "None" for constraints means we can always rely on
`btTypedConstraint` pointers to be valid (current exception is the
"Contact" constraint which is not supported but was intended for future
contact point info. Might get removed). Without null pointers it is much
easier to ensure constraint data is stored persistently, especially when
the type or target bodies are changed.