Revert "Fix #123327: Crash when operator with properties opens render window"
This reverts commit 5f102ad445e003501e41d9f76742154c3408c525.

Apparently this was on the backport list by mistake, a (safer) alternative was commited in 174f8de75e

This 5f102ad445e0 caused #126392 (which is still present in main), we dont want to release 4.2.1 with this though.

Then the current issue only needs to be fixed for 4.3 (main), and we can safely wait for @JulianEisel to be back to do so.