Compositor: Free compositor when context changes
This patch frees the scene compositors if the compositor settings like
precision and execution device change. That's because persistent data
like cached resources are computed for that specific precision and
execution device, and they need to be recomputed if either of those
change. It is just easier to recreate the entire context than to update
the resources that needs updating, so that's what we do.