WIP: VSE: Draw text effect character by character
This PR changes how effect strip draws text. It builds cache, where
position and width of each character is stored. This is necessary for
efficient implementation of text editing in preview

In addition, position and width of each line is stored. This allows to
implement proper text alignment feature, instead of existing anchoring.

This cache is implemented on VSE side, and takes over 2 aspects of text
rendering:
 - Positioning of individual characters (no changes compared to main)
 - Word wrapping (currently buggy)
This approach seems to be used in 3D viewport as well.

TODO:
Anchoring feature may be part of this PR, but can be also committed
separately