Nodes: Singular Value Decomposition for 3x3 matrices.
Singular Value Decomposition (SVD) yields a representation of a 3x3
matrix as two rotations and a scale: M = U * S * V. This is a useful
representation and part of Principal Component Analysis (PCA).
Having this as a separate math node allows the implementation of PCA
as a node group with only a smaller math node at the end.
The covariance matrix can be constructed as a node group and its SVD
yields the principal components and the orientation of maximum variance.