Removed the center-of-mass attribute entirely.
I misunderstood the meaning of `getCenterOfMassTransform` and similar
API functions: these don't change the CoM relative to the shape, but
are just another way of saying "move the rigid body center".

To change the center of mass actually means offsetting the collision
shape itself, for example with a CompoundShape. The CoM of a body is
always at (0,0,0) in local coordinates.