Fix #126412: Metal default alpha value
Metal doesn't support RGB textures and the backend converts them to RGBA textures.
During the conversion missing RGB components should be set to 0 and missing A
component should be set to 1.

In the current implementation this was not the case and A components where also set
to 0.

PR should be backported to 4.2