Fix #126541: EnumProperty dynamic enum items tooltips are broken
Getting the tooltip gets a different codepath depending on whether the
mouse id over the "collapsed" enum button or over one of the buttons of
the expanded dropdown in `ui_tooltip_data_from_button_or_extra_icon`.
The later will use will use the `EnumPropertyItem` `description`, the
former will use the `UIBut` `tip`.
I have checked that the "expanded" buttons get created with the proper
tip (and keep the proper tip if the enum is defined statically), but
somewhere loose them if defined dynamically.

Some digging has been done to why this is (do we just get the wrong
button in `ui_but_tooltip_init`? Do the individual buttons from
`ui_popup_block_create` > `ui_block_func_POPUP` > `ui_def_but_rna__menu`
> `uiDefButI` get the correct tip (yes, they do)?), but this has not
lead to a solution.

However, taking a step back, we can get actually get an
`EnumPropertyItem` even for the buttons in an expanded enum popup and go
fro there. A similar thing is done in
`ui_but_event_property_operator_string` for
example.

So to resolve, get `EnumPropertyItem` via the parent button that spawned
the popup and its info.
This behavior has been added to `UI_but_rna_enum_item_get`.

NOTE: if the real reason for the corruption of the button (tip) data is
found, that would probably be preferred, this bandaid should be fine as
well though?