Better error handling in case of gradients in SVG import.
Gradients cannot be represented well in Grease Pencil. GPv2 was ignoring
this case and simply "interpreting" the gradient pointer as a color,
which is no good. GPv3 was simply outputting black.

To approximate the gradient the average color of all stops is now used.
Additionally a warning is generated when gradients are encountered.