Fix BPY creating an object of the XYZ type to store the SRNA info for the XYZ type.
This is not necessary, a basic `bpy_struct` object is enough to store
the SRNA type definition of a python version of a RNA struct.

Before this commit:
```python
>>> type(C.object.bl_rna)
<class 'bpy_types.Object'>
```

After this commit:
```python
>>> type(C.object.bl_rna)
<class 'bpy.types.Struct'>
```

While this seems harmless in current code, this does break quite badly
in the WIP refactor to support non-trivial `PointerRNA` and related
structs in BPY (see !124790).