Refactor: Sculpt/Paint: Rename brush "tool" to brush "type"
The term "tool" is historic from before the actual tool system got
introduced. Since then the term was already a bit confusing, because it
wasn't directly related to the tool system, but there was still some
relationship between the two. Now brushes and their types are decoupled
much more from the tool system, with a single "Brush" tool supporting
all kinds of brushes (draw, grab, cloth, smooth, ...).

For a much more clear terminology, use "brush type" instead of "tool".

For #126032 we need to write the brush type to the asset metadata (done
in !124618), so we can filter brushes based on the type (so the grease
pencil eraser tool only shows eraser brushes, for example). I'd like to
use future proof names for that to avoid versioning in future, so I'd
rather do the full naming change now.

Included changes:
- Add new RNA properties instead of renaming existing ones. These seem
  commonly accessed from BPY so better keep the legacy names available
  and remove them for 5.0, see #124201.
- Rename DNA members, enumerators, enums,
- Update comments, function names, etc.