Fix: crash when loading layered animation with channel groups
The issue was that the forward compatibility code was setting the
group's channels listbase to point at the forward-compatible fcurve
listbase, but the channels listbase wasn't getting cleared when loading
it as a layered action.  Additionally, the pointers in the group's
channels listbase were just being ignored on load because they aren't
relevant to groups on layered actions, which means they weren't getting
remapped to the new correct memory addresses.

The end result was that after loading the group's channel list base
would be non-null and pointing to invalid memory, and there are some
code paths that try to use that listbase, resulting in a segfault.

This commit fixes the issue by ensuring that those listbases are cleared
when loading channel groups as part of a layered action. Additionally,
we now also clear them after they're set during forward-compatibly
saving, the lack of which wasn't causing problems but was nevertheless
incorrect.