Fix #126808: Crash in sculpt normals update adter topology change
Currently the sculpt BVH tree is cleared at the beginning of depsgraph
reevaluation after a topology changing operation. However, at that point
the BVH and SculptSession data are out of sync with the new mesh.
Before 5fdf7deb57e809dea2d7 that was fine because the PBVH had all the
data necessary for normal recomputation owned locally, but now we
reference pre-existing external data for that.

It's simpler after a topology change to just clear the PBVH-- it doesn't
contain any useful data anyway. In the near future we should move the
ownership to the original mesh, which will make this happen by default
when the mesh topology changes.