Fix #126783: Anim: First key on ID will select previous Slot
Fix an issue that was caused by common code used for both these flows:

- 'ensure a slot exists for an ID so keys for it can be inserted' and
- 'assign an Action to an ID'

Both flows search for suitable slots for the ID, but should do so in
subtly different ways. Most importantly, the first case should _not_
search a slot by name (when an Action is already assigned but not a
slot). This is relevant in the following flow:

- Assign an Action to an Object. This auto-selects the most appropriate
  slot (by name).
- Un-assign the slot (for whatever reason it's not desired).
- Insert a key.
- This should create a new slot, and not re-assign the slot just
  un-assigned above.