Fix #126865: Added movie has incorrect length
Commit 8903205dd9 fixed incorrect math, in stream length setting part of
the code, which seemed to force most video files to use safer, but more
complicated codepath.

Even though the math was corrected, the logic was not great. It tried to
use container duration, only stream duration is not significantly (4x!)
longer. This is quite crude.

Instead, check if there is any significant difference between stream and
container durations.

-----

Note for review: I tried to keep the significance amount about the same as currently in main. However, I would rather make this either more sensitive, or better yet absolute (question is, where is the threshold). But this code is a mess.

For years, `stream_sec` was presumably quite large number, as long as `frame_rate.den != 0 && pFormatCtx->duration > 0` is true, `stream_sec > 4.0 * container_sec` would likely be true as well before 8903205dd9. Therefore why even do this check? I think, that we could immediatelly proceed to use `Fall back to manually estimating the video stream duration.` part of the code.

A very sidenote, I think, that this code could be offloaded to function(s). I've had to reference code by comments and statements...