Fix #126860: Outliner Delete skips selected child collections
`collection_collect_data_to_edit` blocks the tree_traverse once a
valid selected collection is found. This prevents delete operation to
work when child collection is also selected. To fix this, introduce a
new boolean `is_recursive`. When this is true, callback function will
return `TRAVERSE_CONTINUE` to continue the walk to  children.