Fix #126706: Restore 4.2 Widget Label When Using 4.3 Preferences
Although we keep preferences for different blender versions separated
into per-version folders it is still possible to load Blender 4.2 while
using preferences saved with a later Blender version, for example by
copying the file or symlinking folders. Since we removed the "Widget
Label" text style in 4.3 this would result in having such a text style
set to 0 points and therefore unreadable. This PR checks the userdef
version and if newer than 4.2 it checks for this situation and sets
defaults.