UI Experiment: Docking Drag Mouse Cursor
Just a test of showing the default mouse cursor while docking and
simultaneously showing a smaller version of the document with area
cursor. This is mostly an excuse for Pablo to take a look and give
some consideration on the size of the drag image, mouse cursors, and
the best way to give feedback for this process.