Cleanup: Standardize `StrokeCache` symmetry variable naming
Prior to this commit, variables inside `StrokeCache` used either a
`true_` prefix to indicate that the relevant position / normal /
displacement was not modified by symmetry, or a `_symmetry` / `_symm`
suffix to indicate that it was.

This commit changes all such variables to have a `_symmetry` suffix, as
that has fewer connotations than the `true_` prefix.