Fix #126826: Regression in VSE movie rendering
Regression is caused by movie being rendered twice for the same frame
number, but the image is not cached after 5ecb70964e.

The image is rendered first to determine `early_out` in
`seq_render_strip_stack()`, then for the actual image output.

Since VSE cache is not to be used for render job, `ffmpeg_fetchibuf()`
can be optimized to return frame, which is already decoded instead of
re-seeking to last keyframe and decoding all following frames again.