Fix: channel groups in a duplicated channel bag still pointed to old bag
In layered actions, the channel groups in a channel bag hold a
non-owning pointer to the channel bag they belong to, to allow looking
up the fcurves in the group. However, when a duplicate was made of a
channel bag, those pointers in the (also duplicated) channel groups
weren't updated to point at it, and thus were still pointing at the
original channel bag.

This commit adds the code to properly update those pointers when
duplicating channel bags.