Add option to control parallel level for HIP-RT kernels
The default value is 1 which has no effect on the compilation time
compared to the state prior to this change.

With the value of 4 the setup which is similar to the buildbot the
compile time for the new kernels is lowered from 30+ to about min,
which is more acceptable. There does not seep to be huge memory
penalty, so the value might potentially be increased further if
needed.