Fix: layered action forward compat data isn't cleared properly
The issue was that the prev/next listbase pointers were getting cleared
using a `LISTBASE_FOREACH` loop, and thus the loop never progressed past
the first item.

This fixes the issue by switching these loops to use
`LISTBASE_FOREACH_MUTABLE`.

This also adds a clarifying comment to the assert that caught this
issue, explaining that the condition the assert guards against is itself
benign, and is just a sanity check.