Fix: range selecting any animation channels always selects slots
The issue was that slots are using `templated_selection_state_update()`
for their basic selection logic, which is used in the same places as the
`ACHANNEL_SET_FLAG` macro and was written to do the same job. However,
the logic was not properly reproduced.

Specifically, `ACHANNEL_SET_FLAG` actually *deselects* when passed
the `ACHANNEL_SETFLAG_EXTEND_RANGE` or `ACHANNEL_SETFLAG_TOGGLE`
selection mode flags. This makes no sense at all, but it turns out to
be functionally important in the case of `ACHANNEL_SETFLAG_EXTEND_RANGE`
because of how it's used, resulting in this bug.

This fixes the issue by changing `templated_selection_state_update()` to
exactly match the logic of `ACHANNEL_SET_FLAG`, in all its nonsensical
glory. I've also added a comment explaining the situation, so it's clear
why the logic is so odd.