Fix #126901: Inconsistent particle object starting velocity behavior
Spotted the following issues with current code [which evaluates the
emitter at the particle birth time (can be a fraction of a frame) and
compares this to the previous (unborn) particle state coordinate]:

- [1] it is not easy to tell at which exact time this last particle
state coordinate has been saved, thus getting a reliable time delta cant
be determined (and therefor also not the velocity).
-- I have tried scaling the time delta by just the fractional part, but
this also gives wrong results
- [2] code also did not take into account that we can have multiple
integration steps, so with more steps, particles would actually get
faster...

So to resolve [1], we dont rely on the particle state coordinate
anymore, but instead evaluate the emitter again at an earlier point in
time to get a precise matrix. With that, we can get "true" object motion
(at the cost of ... well another evaluation with
`BKE_object_where_is_calc_time`)

To resolve [2], we just have to multiply the time delta with the
`ParticleKey` time

[2] could be split into a separate commit.

NOTE: there might be other ways getting this to work properly (possibly
still using the particle state coordinate and thus avoiding the extra
evaluation), just posting this PR if someone needs it (I do think
correctness beats performance in this case -- that feature is barely
usable otherwise). And yes, I know this this EOL and I should be looking
at something else instead :)