FFmpeg: Simplify code for determining video frame count
Original code was quite unorganized and not as easy to follow through.

There were basically 3 code paths:
1. Division of AVStream duration in seconds by its timebase
2. Division of AVFormatContext duration by its timebase, with possible
   AV offset compensation
3. Simply using frame count from AVStream

Finally there was possibility of ending up with duration of 0 in
specific case where `pFormatCtx->duration == AV_NOPTS_VALUE`, but I did
not find any test case for this. (TODO: did not look at blame yet)

During investigation for PR #126866 I have concluded, that code before
commit 8903205dd9 that compared duration of stream and container was
incorrectly always true, but it resulted in correct behavior for about
4 years. Code was reorganized with that assumption, so above listed code
paths are in order of possible execution.

Further, statement `frame_rate.den != 0` was removed from condition to
code path 1 and 2. This was done, because it does not make sense, it
seems to be based on file characteristic with previously incorrect
behavior, but the characteristic was unrelated to the issue, and
finally, it was not justified anywhere.

-----

TODO test if codepath 1 and 2 should share the condition, because that
does not make sense either.