Compositor: UI: remove preview for viewer and composite nodes
In geometry nodes, the viewer gets activated by clicking on the preview icon. For the compositor, preview is enabled either using the image viewer or the backdrop. Another difference is that all nodes can be previewed in the compositor. Which makes the preview icon on the composite and viewer node redundant.