Fix slight thumbnail "swimming" issues when zoomed in a lot
When zoomed in a lot, the thumbnails were slightly "swimming" to the
sides when changing the zoom level slightly, due to them displaying
integer thumbnail image pixels over the horizontal thumbnail area.
But we don't have to do that, UVs can refer to fractions of a pixel
just fine, so switch thumbnail horizontal "crop" values to be floats.