Fix: layered-action channel groups not shown in action editor
The issue was that we were misinterpreting the `ANIMFILTER_FCURVESONLY`
flag in the animation filtering code. We were taking it to mean that
literally only actual fcurve channels should be included, but in fact
it should include the channel groups for those fcurve channels as well.
The graph editor code (but not the dopesheet code) depended on this
interpretation to function correctly.

This fixes the issue by not excluding channel groups even when that
flag is set. This also updates the documentation of that flag to clarify
things.