Fix #125791: restore authored filepaths on "generated" images
When an image's type is switched to Generated, the first char of the
filepath is changed to '\0', while the rest of the path is left
untouched. IF you switch back to Single Image, or one of the other
types, the authored image file path fials to load because the path is
now invalid.

This change adds a variable to cache off the first character before
overwriting it. The character is restored when a texture switches away
from Generated.